selinux: Remove unnecessary anonymous python blocks
libselinux was attempting to ensure ARCH was set to i386 for any i*86
platform.  Replaced the existing code with a simpler construct that
accomplishes the same goal.

A similar anonymous python block was being used in policycoreutils to
identify an optional dependency on libcap-ng and libcgroup.  Also replaced
with a simpler construct.  The newest policycoreutils depends on them both
anyway in the current configuration.

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>

Added _git versions.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>