audit: Add audit.h from target libc-headers.
The executables in lib/, which are named as gen_*_h, will run on
the hosts to create *_tables.h/*tabs.h header files for the
targets.
In some old hosts, build will fail because audit.h in the old
linux-libc-headers (<= 2.6.29) has a incomplete netlink message
list for the audit system.

Signed-off-by: Xin Ouyang <Xin.Ouyang@windriver.com>