audit: Add .h files from target libc-headers.
Just like audit_2.1.3.bb.

The executables in lib/, which are named as gen_*_h, will run on
the hosts to create *_tables.h/*tabs.h header files for the
targets.
In some old hosts, build will fail because .h files in the old
linux-libc-headers (<= 2.6.29) has incomplete DEFINE lists for
the audit system.

Signed-off-by: Xin Ouyang <Xin.Ouyang@windriver.com>