audit: use generated headers for cross compiling
Default audit Makefile will generate native executables in lib/ and
auparse/, which are named as gen_*_h and run on the hosts to create
*_tables.h/*tabs.h header files for the targets.

This is inappropriate for our cross compiling because they need
linux-libc-headers from the host.

Even worse, on some old hosts, build will fail because some .h files
in the old linux-libc-headers (<= 2.6.29) has incomplete DEFINE lists
for the audit system.

So add *tables.h/*tabs.h header files which are generated from
linux-libc-headers-3.4, and do not generate and run those native
executables.

Signed-off-by: Xin Ouyang <Xin.Ouyang@windriver.com>