libpcre_8.34.bbappend: remove PRINC warning
The libpcre has been upgrade to 8.34 in oe-core, but since we were still
using PRINC until now, we'll need to keep it around (in a PR form), so set
it to something suitably large that it's unlikely to break anyone's
package feed and so that it shows it's clearly an exception case.
Obviously this is just a staging activity until the next update when we
don't include anything of the sort.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>
Signed-off-by: Joe MacDonald <joe@deserted.net>