libselinux: Mount procfs before checking /proc/filesystems
In the case where the SELinux security module is not loaded in the
kernel and it's early enough in the boot process that /proc has not yet
been mounted, selinuxfs_exists() will incorrectly return 1, and
selinux_init_load_policy() will print a message like this to the
console:

Mount failed for selinuxfs on /sys/fs/selinux:  No such file or directory

To fix this, mount the procfs before attempting to open /proc/filesystems, and
unmount it when done if it was initially not mounted.  This is the same
thing that selinux_init_load_policy() does when reading /proc/cmdline.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-CAR-ID: 523015
Natinst-ReviewBoard-ID: 94878