refpolicy: remove virtual prefix for runtime providers
In keeping with the approach of only providing a single default policy at
runtime, we were originally using a virtual/refpolicy dependency and
filling it with one of our specific refpolicy implementations.  This works
well enough for some package systems, but fails for others (specifically
deb, possibly more).

Since the intent was to only have one present in the default image anyway,
we'll just throw out the 'virtual/' part of the RPROVIDES and related
dependencies across the board.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>