audit: fix pkgconfigdir
Because 'libdir' is set with ${base_libdir} in recipe file of audit,
package config files(.pc) are installed to ${base_libdir}/pkgconfig that
variable pkgconfigdir is set with ${libdir}/pkgconfig in makefile.

Set pkgconfigdir directly to install .pc files to right directory.

And remove setting of FILES_${PN}-dev which has been done in
bitbake.conf in oe-core.

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>