packagegroup-selinux-minimal: add selinux-init
When trying to build a minimal image (eg: without python), the default
user on autologin is not mapped to the intended user/role/domain:

  # id -Z
  system_u:system_r:kernel_t:s0

And the following error is displayed on autologin:

  Unable to get valid context for <user>

While on an image built with the core-selinux packagegroup:

  # id -Z
  unconfined_u:unconfined_r:unconfined_t:s0-s0:c0.c1023

Adding selinux-init to the minimal package list fixes the issue.
This package does not seem to bring along additional dependencies.

Signed-off-by: Luca Boccassi <bluca@debian.org>
Signed-off-by: Joe MacDonald <joe@deserted.net>